revise commit 432b205e3aa84436c3685d2ca6f0dc5a766177c3

+++ include/ir/id_manager/IDManager.h
@@ -452,7 +452,7 @@ typedef _IDManager<izenelib::util::UString, uint128_t, uint32_t,
                    HashIDGenerator<izenelib::util::UString, uint32_t>,
                    EmptyIDStorage<izenelib::util::UString, uint32_t>,
                    UniqueIDGenerator<uint128_t, uint32_t>,
-                   EmptyIDStorage<uint32_t, uint32_t> > IDManager;
+                   EmptyIDStorage<uint128_t, uint32_t> > IDManager;
 #endif
 
 } // end - namespace idmanager
